Add computing cepstrum of the whole sound
