+++ src/utils/sound_utils.py
@@ -1,4 +1,7 @@
 import numpy as np
+import scipy.signal as sig
+import math
+import numpy as np
 
 from src.classes.containers import SoundWave
 from src.conf import config
@@ -22,6 +25,37 @@ def preemphasis(sound_wave):
     sound_wave.samples = np.append(sound_wave.samples[0], sound_wave.samples[1:] - pre * sound_wave.samples[:-1])
 
     return sound_wave
+
+
+def downsample(sound_wave):
+    sampling_frequency = config.analysis['sampling_frequency']
+
+    # down and up factor
+    last_common_multiple = (sound_wave.fs * sampling_frequency) / math.gcd(sound_wave.fs, sampling_frequency)
+    upsample_factor = int(last_common_multiple // sound_wave.fs)
+    downsample_factor = int(last_common_multiple // sampling_frequency)
+
+    # upsampling
+    audio_up = np.zeros(sound_wave.length() * upsample_factor)
+    audio_up[upsample_factor // 2::upsample_factor] = sound_wave.samples
+
+    # filtering
+    alias_filter = sig.firwin(301, cutoff=sampling_frequency / 2, fs=sound_wave.fs * upsample_factor)
+    audio_up = downsample_factor * sig.filtfilt(alias_filter, 1, audio_up)
+
+    # downsampling
+    audio_down = audio_up[downsample_factor // 2::downsample_factor]
+    sound_wave.samples = audio_down
+    sound_wave.fs = sampling_frequency
+
+    return sound_wave
+
+
+def remove_dc_offset(sound_wave):
+    sound_wave.samples = sound_wave.samples - np.mean(sound_wave.samples)
+
+    return sound_wave
+
 
 
 def detect_speech(sound_wave):
